pyproject.toml
[tool.poetry]
name = "personal-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.4"
pandas = "^2.2.2"
loguru = "^0.7.2"
llama-index = "^0.11.2"
llama-index-llms-ollama = "^0.3.0"
llama-index-embeddings-huggingface = "^0.3.1"
maturin = "0.8.2"
pyarrow = "^17.0.0"
matplotlib = "^3.9.2"
python-dotenv = "^1.0.1"
mlflow = "^2.16.0"
llama-index-llms-together = "^0.2.0"
llama-index-embeddings-ollama = "^0.3.0"
arize-phoenix = "^4.33.0"
llama-index-callbacks-arize-phoenix = "^0.2.1"
ipywidgets = "^8.1.5"
transformers = "4.43.4"
llama-index-vector-stores-qdrant = "^0.3.0"
llama-index-postprocessor-flag-embedding-reranker = "^0.2.0"
llama-index-retrievers-bm25 = "^0.3.0"
llama-index-llms-gemini = "^0.3.5"
flagembedding = {git = "https://github.com/FlagOpen/FlagEmbedding.git"}
spacy = "^3.7.6"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.2.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
